tests/wpt/tests/css/css-box/margin-trim/computed-margin-values/grid-block-end-column-auto-flow.html

<!DOCTYPE html>
<html>
<head>
<link rel="author" href="mailto:sammy.gill@apple.com">
<link rel="help" href="https://drafts.csswg.org/css-box-4/#margin-trim-grid">
<meta name="assert" content="Trimmed block-end margins for grid items should be reflected in computed style">
</head>
<style>
grid {
    display: grid;
    width: min-content;
    grid-auto-flow: column;
    border: 1px solid black;
    grid-template-rows: auto auto auto;
    margin-trim: block-end;
}
item {
    display: block;
    width: 50px;
    height: 50px;
    margin-block-end: 10px;
}
.span-four {
    grid-row: span 4;
}
item:nth-child(1) {
    grid-row: 1;
    background-color: green;
}
item:nth-child(2) {
    grid-row: 2;
    background-color: blue;
}
item:nth-child(3) {
    grid-row: 3;
    background-color: purple;
}
item:nth-child(4) {
    background-color:burlywood;
}
item:nth-child(5) {
    background-color: grey;
    grid-row: 4;
}
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<body onload="checkLayout('grid > item')">
    <div id="target">
    <grid>
        <item data-expected-margin-bottom="10"></item>
        <item data-expected-margin-bottom="10"></item>
        <item data-expected-margin-bottom="10"></item>
        <item data-expected-margin-bottom="0" class="span-four"></item>
        <item data-expected-margin-bottom="0"></item>
    </grid>
    </div>
</body>
</html>
